% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalpDataPreprocessing-package.R,
%   R/metalpDataPreprocessing.R
\docType{package}
\name{metalpDataPreprocessing}
\alias{metalpDataPreprocessing}
\title{METALP Sensor Data Preprocessing}
\usage{
metalpDataPreprocessing()
}
\description{
Use this addin to perform all the data preprocessing for the METALP sensors data

Run this addin to open a window to run some preprocessing tasks for the
METALP Sensor data.
}
