% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalpDataPreprocessing.R
\name{metalpDataPreprocessing}
\alias{metalpDataPreprocessing}
\title{Sensor Data Preprocessing for METALP Data}
\usage{
metalpDataPreprocessing()
}
\description{
Run this addin to open a window to run some preprocessing tasks for the
METALP Sensor data.
}
